requirements.txt
discord.py
gtts
python-dotenv
virtualenv